Implemented Transform tracks, Pose class, Clip class
Implemented the transform track class which is a higher level than the normal track class. Pose class hold the current data for a frame of an animation for all bones. The clip class holds all information about a single animation clip. added functions to the loader to load in the rest pose for a model and to load in all the animation clips from a file